fix rename that broke during merge
